edited

+++ print_binary.c
@@ -1,7 +1,7 @@
 #include "main.h"
 
 /**
- * op_int_helper - prints 
+ * op_int_helper - prints
  * @n: integer
  *
  * Return: the length of @n
@@ -19,18 +19,18 @@ int op_binary_helper(long int n)
 }
 
 /**
- * op_int - prints 
+ * op_int - prints
  * @vl: varasic list of arguments
  *
  * Return: the length of the argument
  */
 int op_binary(va_list vl)
 {
-	int len = 0;
-	long int n = va_arg(vl, int);
+        int len = 0;
+        unsigned int n = va_arg(vl, unsigned int);
 
-	if (n / 2 >= 0)
-		len += op_binary_helper(n);
+        if (n / 2 >= 0)
+                len += op_binary_helper(n);
 
-	return (len);
+        return (len);
 }
